+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## **0.0.3** - *2020-03-07*
+* accept `Buffer` and `Uint8Array` as a source;
 * readme fixed;
 
 ## **0.0.2** - *2020-03-07*